Refactor Dockerfiles so app images build in the full project context
This is so that all packages can make use of other packages in this
monorepository. If the docker context is local to the project then
packages can't be found. Using the full project as the context has its
downsides though; we have to manually copy only those packages that are
relevent into images so that the image doesn't have to include all
dependencies and source code for the entire project. This means that we
need to manage local dependenceis in the docker files.

This change also optimises the images a bit by using mult-stage builds
for the webappa and allows all dockerfiles to optimially use docker layer
caching by installing dependencies before copying source code so that a
source code change doesn't alter the layer with the dependencies
installed.