Fix docker builds
We weren't copying dependent modules from the build image, this means
that when the build files don't already exist in the docker context
(i.e. when the module hasn't been built locally) then the build files
don't exit and the module cannot be found. This change corrects this by
copying the build files from the build image as it should be.