always retry mongo-on-update on failure
while I was testing 1.6.0, I noticed my staging environment not working. When trying to reset and seed the environment, I noticed reset not working as there was no replica set initialised in MongoDB. What had happened was this service failing for whatever reason. Kicking it back up and then doing a reset fixed the issue