MOSIP interoperability for 1.9 (#1716)
* add initial mosip interoperability for Farajaland 1.9

* bump mosip api version

* use new version for mosip repo

* upgrade toolkit

* add verification status for mother

* update env variables as documented

* update the translation to be more realistic

description todo still

* update description translation

* move mosip interop url to be changeable with env vars

* add verification for father and informant

* fix failed translations

* fix typo with mutating declaration instead of updatedFields

* add deceased verification

* update toolkit according to core

* update text to status as per toolkit new version

* add nid field for child

* bump mosip package

* add NID to birth certificate

* fix aggregate not needing 2nd param anymore

* chore: restore mosip fields in v2 form

* update mosip api to 1.9.0-beta.4

* sync with latest form changes

* remove old mosip api references

* bump to beta.6

* bump mosip api version

* fix ui not waiting for external validation in e2es

* fix: 4.5.2, 3.7.2

* fix: function name typo

* ensure external validation is empty in 1.4.2 2.2.6 5.5

* ensure external validation is empty in 8.3.8

* chore: disable 2FA on v19-beta-prod

This reverts commit e65363cbb46d820779eb52364b2dea9d769d5ad4.

* chore: enable 2FA on v19-beta-prod

This reverts commit b85245fb8ccd10038ac194a10fff05ed9df66273.

* try if ensuring external validation fixes form state test

* align farajaland custom-event/handler with countryconfig-mosip

* fix: remove registrationNumber validation for external ID providers

When an external ID provider (e.g., MOSIP) is integrated, the
registrationNumber may not be immediately available after a
registration is submitted.By the time the script logs in and
navigates to the record audit view, the registrationNumber might
already exist in the UI, but it’s not reliably accessible as a
reference value in this context.

Removing the strict check ensures smoother flow and avoids false
negatives in such cases.

* copy mosip.ts properly over from opencrvs-countryconfig-mosip

* fix(correction-summary): remove registrationNumber validation for external ID providers
When an external ID provider (e.g., MOSIP) is integrated, the
registrationNumber may not be immediately available after a
registration is submitted.By the time the script logs in and
navigates to the record audit view, the registrationNumber might
already exist in the UI, but it’s not reliably accessible as a
reference value in this context.

Removing the strict check ensures smoother flow and avoids false
negatives in such cases.

---------

Co-authored-by: Md. Ashikul Alam <ashikul.alam@dsinnovators.com>
Co-authored-by: Md. Ashikul Alam <32668488+Nil20@users.noreply.github.com>
Co-authored-by: tahmidrahman-dsi <tahmid.rahman@dsinnovators.com>
Co-authored-by: jamil314 <jamil31415926@gmail.com>